reafctor: convert stack to slice mutex
